Logs/metagpt/part3-2/instruction39/algo_trading_backend_python/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic==2.7.1
python-jose==3.3.0
requests==2.31.0
passlib[bcrypt]==1.7.4
alembic==1.13.1
loguru==0.7.2